Add custom hook
